+++ src/Utils/Utils.fs
@@ -188,3 +188,29 @@ module Logging =
                     .AddConsole(fun c -> c.LogToStandardErrorThreshold <- LogLevel.Error)
                 |> ignore
             )
+
+[<RequireQualifiedAccess>]
+module Serialize =
+    module private Json =
+        open Newtonsoft.Json
+        open Newtonsoft.Json.Serialization
+
+        let private options () =
+            JsonSerializerSettings (
+                ContractResolver =
+                    DefaultContractResolver (
+                        NamingStrategy = SnakeCaseNamingStrategy()
+                    )
+            )
+
+        let serialize obj =
+            JsonConvert.SerializeObject (obj, options())
+
+        let serializePretty obj =
+            let options = options()
+            options.Formatting <- Formatting.Indented
+
+            JsonConvert.SerializeObject (obj, options)
+
+    let toJsonPretty: obj -> string = Json.serializePretty
+    let toJson: obj -> string = Json.serialize